Adjust elasticsearch to favour recent results.
This applies a gaussian decay function to the
scores, starting when docs are 5 days old,
such that at 30 days old scores are 60% (0.6)
of what they'd normally be.

This will ensure we favour recent results without
biasing too heavily.

This will help to ensure that more recent, revelant
docs show up in search alerts, otherwise we
tend to have older docs in the top 20 results (simply
because there are more of them).

Fixes #54.