Include committee and content date as dimensions in GA
This allows us to see which committee content is popular,
and track pageviews by content date (eg. which 2016 content
is the most popular).

We could also do this by adding the committee id and/or
date to the URL, but that would break old metrics
and we're hoping this is simpler.