Page through to all current bills to ensure mods to old bills show
The API sorts bills by year. There are currently only around 30 current bills but
if this grows we want to keep seeing modifications to bills from earlier years
come up on the home page. This should only become slow when there are a massive number
of current bills and caching can easily deal with that if it becomes relevant